[DataModel] separated calculated fields to a new module
If you use a full import (e.g. from kitty.model.low_level.field import
Hash), this will break your code, and in general you should use a more
generic import, e.g. from kitty.model.low_level import Hash.
