+++ docs/source/kitty.model.low_level.calculated.rst
@@ -1,0 +1,7 @@
+kitty.model.low_level.calculated module
+=======================================
+
+.. automodule:: kitty.model.low_level.calculated
+    :members:
+    :undoc-members:
+    :show-inheritance:
