+++ kitty/model/low_level/__init__.py
@@ -19,9 +19,10 @@
 This package contains the low level data model, which represents the structure
 of specific messages in the fuzzed protocol.
 '''
+from kitty.model.low_level.aliases import *
+from kitty.model.low_level.calculated import *
 from kitty.model.low_level.condition import *
-from kitty.model.low_level.field import *
 from kitty.model.low_level.container import *
 from kitty.model.low_level.encoder import *
-from kitty.model.low_level.aliases import *
+from kitty.model.low_level.field import *
 from kitty.model.low_level.mutated_field import *
