+++ kitty/model/low_level/aliases.py
@@ -89,7 +89,8 @@ Functions that provide simpler field creation API and name aliasing for convenie
     | '<='      | LesserEqual, AtMost   |
     +-----------+-----------------------+
 '''
-from kitty.model.low_level.field import BitField, Size, Hash
+from kitty.model.low_level.field import BitField
+from kitty.model.low_level.calculated import Size, Hash
 from kitty.model.low_level.encoder import ENC_INT_DEFAULT, ENC_INT_BE, ENC_INT_LE, ENC_STR_DEFAULT
 from kitty.model.low_level.condition import Compare
 
@@ -97,66 +98,82 @@ from kitty.model.low_level.condition import Compare
 # ################### Field Aliases ####################
 
 def UInt8(value, min_value=None, max_value=None, encoder=ENC_INT_DEFAULT, fuzzable=True, name=None):
+    '''Unsigned 8-bit field'''
     return BitField(value, 8, signed=False, min_value=min_value, max_value=max_value, encoder=encoder, fuzzable=fuzzable, name=name)
 
 
 def UInt16(value, min_value=None, max_value=None, encoder=ENC_INT_DEFAULT, fuzzable=True, name=None):
+    '''Unsigned 16-bit field'''
     return BitField(value, 16, signed=False, min_value=min_value, max_value=max_value, encoder=encoder, fuzzable=fuzzable, name=name)
 
 
 def UInt32(value, min_value=None, max_value=None, encoder=ENC_INT_DEFAULT, fuzzable=True, name=None):
+    '''Unsigned 32-bit field'''
     return BitField(value, 32, signed=False, min_value=min_value, max_value=max_value, encoder=encoder, fuzzable=fuzzable, name=name)
 
 
 def UInt64(value, min_value=None, max_value=None, encoder=ENC_INT_DEFAULT, fuzzable=True, name=None):
+    '''Unsigned 64-bit field'''
     return BitField(value, 64, signed=False, min_value=min_value, max_value=max_value, encoder=encoder, fuzzable=fuzzable, name=name)
 
 
 def SInt8(value, min_value=None, max_value=None, encoder=ENC_INT_DEFAULT, fuzzable=True, name=None):
+    '''Signed 8-bit field'''
     return BitField(value, 8, signed=True, min_value=min_value, max_value=max_value, encoder=encoder, fuzzable=fuzzable, name=name)
 
 
 def SInt16(value, min_value=None, max_value=None, encoder=ENC_INT_DEFAULT, fuzzable=True, name=None):
+    '''Signed 16-bit field'''
     return BitField(value, 16, signed=True, min_value=min_value, max_value=max_value, encoder=encoder, fuzzable=fuzzable, name=name)
 
 
 def SInt32(value, min_value=None, max_value=None, encoder=ENC_INT_DEFAULT, fuzzable=True, name=None):
+    '''Signed 32-bit field'''
     return BitField(value, 32, signed=True, min_value=min_value, max_value=max_value, encoder=encoder, fuzzable=fuzzable, name=name)
 
 
 def SInt64(value, min_value=None, max_value=None, encoder=ENC_INT_DEFAULT, fuzzable=True, name=None):
+    '''Signed 64-bit field'''
     return BitField(value, 64, signed=True, min_value=min_value, max_value=max_value, encoder=encoder, fuzzable=fuzzable, name=name)
 
 
 def BE8(value, min_value=None, max_value=None, fuzzable=True, name=None):
+    '''8-bit field, Big endian encoded'''
     return UInt8(value, min_value=min_value, max_value=max_value, encoder=ENC_INT_BE, fuzzable=fuzzable, name=name)
 
 
 def BE16(value, min_value=None, max_value=None, fuzzable=True, name=None):
+    '''16-bit field, Big endian encoded'''
     return UInt16(value, min_value=min_value, max_value=max_value, encoder=ENC_INT_BE, fuzzable=fuzzable, name=name)
 
 
 def BE32(value, min_value=None, max_value=None, fuzzable=True, name=None):
+    '''32-bit field, Big endian encoded'''
     return UInt32(value, min_value=min_value, max_value=max_value, encoder=ENC_INT_BE, fuzzable=fuzzable, name=name)
 
 
 def BE64(value, min_value=None, max_value=None, fuzzable=True, name=None):
+    '''64-bit field, Big endian encoded'''
     return UInt64(value, min_value=min_value, max_value=max_value, encoder=ENC_INT_BE, fuzzable=fuzzable, name=name)
 
 
 def LE8(value, min_value=None, max_value=None, fuzzable=True, name=None):
+    '''8-bit field, Little endian encoded'''
     return UInt8(value, min_value=min_value, max_value=max_value, encoder=ENC_INT_LE, fuzzable=fuzzable, name=name)
 
 
 def LE16(value, min_value=None, max_value=None, fuzzable=True, name=None):
+    '''16-bit field, Little endian encoded'''
     return UInt16(value, min_value=min_value, max_value=max_value, encoder=ENC_INT_LE, fuzzable=fuzzable, name=name)
 
 
 def LE32(value, min_value=None, max_value=None, fuzzable=True, name=None):
+    '''32-bit field, Little endian encoded'''
     return UInt32(value, min_value=min_value, max_value=max_value, encoder=ENC_INT_LE, fuzzable=fuzzable, name=name)
 
 
 def LE64(value, min_value=None, max_value=None, fuzzable=True, name=None):
+    '''64-bit field, Little endian encoded'''
     return UInt64(value, min_value=min_value, max_value=max_value, encoder=ENC_INT_LE, fuzzable=fuzzable, name=name)
 
 U8 = UInt8
@@ -182,30 +199,58 @@ QwordLE = LE64
 # ################### Calculated Field Aliases ####################
 
 def SizeInBytes(sized_field, length, encoder=ENC_INT_DEFAULT, fuzzable=False, name=None):
+    '''
+    :rtype: :class:`~kitty.model.low_level.calculated.Size`
+    :return: Size field, which holds the size in bytes of the sized field
+    '''
     return Size(sized_field=sized_field, length=length, calc_func=lambda x: len(x) / 8, encoder=encoder, fuzzable=fuzzable, name=name)
 
 
 def Md5(depends_on, encoder=ENC_STR_DEFAULT, fuzzable=False, name=None):
+    '''
+    :rtype: :class:`~kitty.model.low_level.calculated.Hash`
+    :return: MD5 hash field
+    '''
     return Hash(depends_on=depends_on, algorithm='md5', encoder=encoder, fuzzable=fuzzable, name=name)
 
 
 def Sha1(depends_on, encoder=ENC_STR_DEFAULT, fuzzable=False, name=None):
+    '''
+    :rtype: :class:`~kitty.model.low_level.calculated.Hash`
+    :return: SHA1 hash field
+    '''
     return Hash(depends_on=depends_on, algorithm='sha1', encoder=encoder, fuzzable=fuzzable, name=name)
 
 
 def Sha224(depends_on, encoder=ENC_STR_DEFAULT, fuzzable=False, name=None):
+    '''
+    :rtype: :class:`~kitty.model.low_level.calculated.Hash`
+    :return: SHA224 hash field
+    '''
     return Hash(depends_on=depends_on, algorithm='sha224', encoder=encoder, fuzzable=fuzzable, name=name)
 
 
 def Sha256(depends_on, encoder=ENC_STR_DEFAULT, fuzzable=False, name=None):
+    '''
+    :rtype: :class:`~kitty.model.low_level.calculated.Hash`
+    :return: SHA256 hash field
+    '''
     return Hash(depends_on=depends_on, algorithm='sha256', encoder=encoder, fuzzable=fuzzable, name=name)
 
 
 def Sha384(depends_on, encoder=ENC_STR_DEFAULT, fuzzable=False, name=None):
+    '''
+    :rtype: :class:`~kitty.model.low_level.calculated.Hash`
+    :return: SHA384 hash field
+    '''
     return Hash(depends_on=depends_on, algorithm='sha384', encoder=encoder, fuzzable=fuzzable, name=name)
 
 
 def Sha512(depends_on, encoder=ENC_STR_DEFAULT, fuzzable=False, name=None):
+    '''
+    :rtype: :class:`~kitty.model.low_level.calculated.Hash`
+    :return: SHA512 hash field
+    '''
     return Hash(depends_on=depends_on, algorithm='sha512', encoder=encoder, fuzzable=fuzzable, name=name)
 
 
@@ -214,6 +259,8 @@ def Sha512(depends_on, encoder=ENC_STR_DEFAULT, fuzzable=False, name=None):
 def Equal(field, comp_value):
     '''
     Condition applies if the value of the field is equal to the comp_value
+
+    :rtype: :class:`~kitty.model.low_level.condition.Compare`
     '''
     return Compare(field, '==', comp_value)
 
@@ -221,6 +268,8 @@ def Equal(field, comp_value):
 def NotEqual(field, comp_value):
     '''
     Condition applies if the value of the field is not equal to the comp_value
+
+    :rtype: :class:`~kitty.model.low_level.condition.Compare`
     '''
     return Compare(field, '!=', comp_value)
 
@@ -228,6 +277,8 @@ def NotEqual(field, comp_value):
 def Greater(field, comp_value):
     '''
     Condition applies if the value of the field is greater than the comp_value
+
+    :rtype: :class:`~kitty.model.low_level.condition.Compare`
     '''
     return Compare(field, '>', comp_value)
 
@@ -235,6 +286,8 @@ def Greater(field, comp_value):
 def GreaterEqual(field, comp_value):
     '''
     Condition applies if the value of the field is greater than or equals to the comp_value
+
+    :rtype: :class:`~kitty.model.low_level.condition.Compare`
     '''
     return Compare(field, '>=', comp_value)
 
@@ -242,6 +295,8 @@ def GreaterEqual(field, comp_value):
 def Lesser(field, comp_value):
     '''
     Condition applies if the value of the field is lesser than the comp_value
+
+    :rtype: :class:`~kitty.model.low_level.condition.Compare`
     '''
     return Compare(field, '<', comp_value)
 
@@ -249,6 +304,8 @@ def Lesser(field, comp_value):
 def LesserEqual(field, comp_value):
     '''
     Condition applies if the value of the field is lesser than or equals to the comp_value
+
+    :rtype: :class:`~kitty.model.low_level.condition.Compare`
     '''
     return Compare(field, '<=', comp_value)
