+++ kitty/model/low_level/field.py
@@ -23,16 +23,13 @@ Each "field" type is a discrete component in the full Template.
 from random import Random
 import copy
 import os
-import types
-import zlib
-import hashlib
 import logging
 from bitstring import Bits
 from kitty.core import KittyObject, KittyException, kassert, khash
 from kitty.model.low_level.encoder import ENC_STR_DEFAULT, StrEncoder
 from kitty.model.low_level.encoder import ENC_INT_DEFAULT, BitFieldEncoder
 from kitty.model.low_level.encoder import ENC_BITS_DEFAULT, BitsEncoder
-from kitty.model.low_level.ll_utils import RenderContext
+
 
 empty_bits = Bits()
 
@@ -865,10 +862,9 @@ class RandomBits(BaseField):
     _encoder_type_ = BitsEncoder
 
     def __init__(
-        self, value, min_length, max_length, unused_bits=0,
-        seed=1235, num_mutations=25, step=None, encoder=ENC_BITS_DEFAULT,
-        fuzzable=True, name=None
-    ):
+            self, value, min_length, max_length, unused_bits=0,
+            seed=1235, num_mutations=25, step=None, encoder=ENC_BITS_DEFAULT,
+            fuzzable=True, name=None):
         '''
         :type value: str
         :param value: default value, the last *unsused_bits* will be removed from the value
@@ -1018,417 +1014,3 @@ class RandomBytes(BaseField):
         '''
         hashed = super(RandomBytes, self).hash()
         return khash(hashed, self._min_length, self._max_length, self._num_mutations, self._step, self._seed)
-
-
-# ################### Calculated fields ####################
-# Fields that are dependant on the values of other
-# fields - Size, Checksum etc.
-# ##########################################################
-
-
-class Calculated(BaseField):
-    '''
-    A base type for fields that are calculated based on other fields
-    '''
-    _encoder_type_ = BitsEncoder
-    _default_value_ = empty_bits
-
-    def __init__(self, depends_on, encoder=ENC_BITS_DEFAULT, fuzzable=True, name=None):
-        '''
-        :param depends_on: (name of) field we depend on
-        :type encoder: :class:`~kitty.model.low_levele.encoder.BitsEncoder`
-        :param encoder: encoder for the field
-        :param fuzzable: is container fuzzable
-        :param name: (unique) name of the container
-        '''
-        self._rendered_field = None
-        super(Calculated, self).__init__(value=self.__class__._default_value_, encoder=encoder, fuzzable=fuzzable, name=name)
-        if isinstance(depends_on, types.StringTypes):
-            self._field_name = depends_on
-            self._field = None
-        elif isinstance(depends_on, BaseField):
-            self._field_name = None
-            self._field = depends_on
-        else:
-            raise KittyException('depends_on parameter (%s) is neither a string nor a valid field' % depends_on)
-
-    def is_default(self):
-        '''
-        Checks if the field is in its default form
-
-        :return: True if field is in default form
-        '''
-        return False
-
-    def _initialize(self):
-        '''
-        We override _initialize, as we want to resolve the field each time
-        '''
-        if self._field_name:
-            self._field = self.resolve_field(self._field_name)
-        if not self._field:
-            raise KittyException('Could not resolve field name %s' % self._field_name)
-
-    def render(self, ctx=None):
-        '''
-        Render the current value into a :class:`bitstring.Bits` object
-
-        :rtype: :class:`bitstring.Bits`
-        :return: the rendered field
-        '''
-        self._initialize()
-        if ctx is None:
-            ctx = RenderContext()
-        #
-        # if we are called from within render, return a dummy object...
-        #
-        if self in ctx:
-            self._current_rendered = self._in_render_value()
-        else:
-            ctx.push(self)
-            self._rendered_field = self._field.render(ctx)
-            self._render()
-            ctx.pop()
-        return self._current_rendered
-
-    def _render(self):
-        raise NotImplementedError('_render should be implemented in subclass')
-
-    def _in_render_value(self):
-        '''
-        This method is called when rendered was called recursively.
-        This means that we are enclosed by the field we depend on.
-        So consider carefully what value are you going to return....
-
-        :rtype: Bits
-        :return: a dummy rendered value
-        '''
-        raise NotImplementedError('_in_render_value should be implemented in subclass (%s)' % type(self).__name__)
-
-    def hash(self):
-        '''
-        :rtype: int
-        :return: hash of the field
-        '''
-        hashed = super(Calculated, self).hash()
-        return khash(hashed, self._field_name)
-
-
-class CalculatedBits(Calculated):
-    '''
-    field that depends on the rendered value of a field, and rendered into Bits() object
-    '''
-    def __init__(self, depends_on, func, encoder=ENC_BITS_DEFAULT, fuzzable=True, name=None):
-        '''
-        :param depends_on: (name of) field we depend on
-        :type encoder: :class:`~kitty.model.low_levele.encoder.BitsEncoder`
-        :param func: function for processing of the dependant data. func(Bits)->Bits
-        :param encoder: encoder for the field
-        :param fuzzable: is container fuzzable
-        :param name: (unique) name of the container
-        '''
-        try:
-            res = func(empty_bits)
-            kassert.is_of_types(res, Bits)
-            self._func = func
-        except:
-            raise KittyException('func should be func(Bits)->Bits')
-        super(CalculatedBits, self).__init__(depends_on=depends_on, encoder=encoder, fuzzable=fuzzable, name=name)
-
-    def _render(self):
-        res = self._func(self._rendered_field)
-        self.set_current_value(res)
-
-
-class Clone(CalculatedBits):
-    '''
-    rendered the same as the field it depends on
-    '''
-    def __init__(self, depends_on, encoder=ENC_BITS_DEFAULT, fuzzable=False, name=None):
-        '''
-        :param depends_on: (name of) field we depend on
-        :type encoder: :class:`~kitty.model.low_levele.encoder.BitsEncoder`
-        :param encoder: encoder for the field (default: ENC_BITS_DEFAULT)
-        :param fuzzable: is container fuzzable
-        :param name: (unique) name of the container
-        '''
-        super(Clone, self).__init__(depends_on=depends_on, func=lambda x: x, encoder=encoder, fuzzable=fuzzable, name=name)
-
-    def _in_render_value(self):
-        return empty_bits
-
-
-class CalculatedStr(Calculated):
-    '''
-    field that depends on the rendered value of a byte-aligned field and rendered to a byte aligned Bits() object
-    '''
-    _encoder_type_ = StrEncoder
-    _default_value_ = ''
-
-    def __init__(self, depends_on, func, encoder=ENC_STR_DEFAULT, fuzzable=False, name=None):
-        '''
-        :param depends_on: (name of) field we depend on
-        :param func: function for processing of the dependant data. func(str)->str
-        :type encoder: :class:`~kitty.model.low_levele.encoder.StrEncoder`
-        :param encoder: encoder for the field (default: ENC_STR_DEFAULT)
-        :param fuzzable: is container fuzzable
-        :param name: (unique) name of the container
-        '''
-        try:
-            res = func('')
-            kassert.is_of_types(res, types.StringTypes)
-            self._func = func
-        except:
-            raise KittyException('func should be func(str)->str')
-        super(CalculatedStr, self).__init__(depends_on=depends_on, encoder=encoder, fuzzable=fuzzable, name=name)
-
-    def _render(self):
-        if len(self._rendered_field) % 8 != 0:
-            raise KittyException('Hashed data should be byte aligned')
-        digest = self._func(self._rendered_field.bytes)
-        self.set_current_value(digest)
-
-
-class Hash(CalculatedStr):
-    '''
-    Hash of a field.
-
-    .. note::
-
-        To make it more convenient, there are multiple aliases for various hashes.
-        Take a look at :mod:`~kitty.model.low_level.aliases`.
-    '''
-    _algos = {
-        'md5': (hashlib.md5, 128),
-        'sha1': (hashlib.sha1, 160),
-        'sha224': (hashlib.sha224, 224),
-        'sha256': (hashlib.sha256, 256),
-        'sha384': (hashlib.sha384, 384),
-        'sha512': (hashlib.sha512, 512),
-    }
-
-    def __init__(self, depends_on, algorithm, encoder=ENC_STR_DEFAULT, fuzzable=False, name=None):
-        '''
-        :param depends_on: (name of) field to be hashed
-        :param algorithm: hash algorithm name (from Hash._algos) or a function to calculate the value of the field. func(str) -> str
-        :type encoder: :class:`~kitty.model.low_levele.encoder.StrEncoder`
-        :param encoder: encoder for the field (default: ENC_STR_DEFAULT)
-        :param fuzzable: is field fuzzable (default: False)
-        :param name: (unique) name of the field (default: None)
-        '''
-        if algorithm in Hash._algos:
-            algo = Hash._algos[algorithm][0]
-
-            def algo_func(x):
-                return algo(x).digest()
-
-            func = algo_func
-            self._hash_length = Hash._algos[algorithm][1]
-        else:
-            try:
-                res = algorithm('')
-                kassert.is_of_types(res, types.StringTypes)
-                func = algorithm
-                self._hash_length = len(res) * 8
-            except:
-                raise KittyException('algorithm should be a func(str)->str or one of the strings %s' % (Hash._algos.keys(),))
-        super(Hash, self).__init__(depends_on=depends_on, func=func, encoder=encoder, fuzzable=fuzzable, name=name)
-
-    def _in_render_value(self):
-        return Bits(self._hash_length)
-
-
-class CalculatedInt(Calculated):
-    '''
-    field that depends on the rendered value of another field and is rendered to (int, length, signed) tuple
-    '''
-
-    def __init__(self, depends_on, bit_field, calc_func, encoder=ENC_BITS_DEFAULT, fuzzable=False, name=None):
-        '''
-        :param depends_on: (name of) field we depend on
-        :param bit_field: a BitField to be used for holding the value
-        :type calc_func: func(bits) -> int
-        :param calc_func: function to calculate the value of the field
-        :type encoder: :class:`~kitty.model.low_level.encoder.BitsEncoder`
-        :param encoder: encoder for the field (default: ENC_BITS_DEFAULT)
-        :param fuzzable: is container fuzzable
-        :param name: (unique) name of the container
-        '''
-        super(CalculatedInt, self).__init__(depends_on=depends_on, encoder=encoder, fuzzable=fuzzable, name=name)
-        self._bit_field = bit_field
-        self._calc_func = calc_func
-        self._first_render = False
-        self._num_mutations = self._bit_field.num_mutations()
-
-    def scan_for_field(self, field_name):
-        '''
-        If the field name is the internal field - return it
-        '''
-        if self._bit_field.get_name() == field_name:
-            return self._bit_field
-        else:
-            return None
-
-    def reset(self):
-        super(CalculatedInt, self).reset()
-        self._first_render = False
-
-    def _calculate_value(self):
-        return self._calc_func(self._rendered_field)
-
-    def _render(self):
-        calculated_value = self._calculate_value()
-        # This code meant for handling overflow...
-        calculated_value = min(calculated_value, self._bit_field._max_value)
-        calculated_value = max(calculated_value, self._bit_field._min_value)
-        if self._mutating():
-            if self._first_render:
-                # mutate applies a mutation function on the default value
-                self._bit_field._default_value = calculated_value
-                self._bit_field.mutate()
-                self._first_render = False
-        else:
-            self._bit_field.set_current_value(calculated_value)
-        self.set_current_value(self._bit_field.render())
-
-    def _mutate(self):
-        self._first_render = True
-
-    def _in_render_value(self):
-        '''
-        :return: a zeroed version of the field, good for some checksums and inclusive lengths
-        '''
-        return Bits(len(self._bit_field.render()))
-
-
-class FieldIntProperty(CalculatedInt):
-    '''
-    Calculate an int value based on some field property.
-    The main difference from :class:`~kitty.model.low_level.field.CalculatedInt`
-    is that it provides the field itself to the calculation function,
-    not its rendered value.
-    '''
-    def __init__(self, depends_on, length, correction=None, encoder=ENC_INT_DEFAULT, fuzzable=False, name=None):
-        '''
-        :param depends_on: (name of) field we depend on
-        :param length: length of the FieldIntProperty field (in bits)
-        :type corrention: int or func(int) -> int
-        :param correction: correction function, or value for the index
-        :type calc_func: func(field) -> int
-        :param calc_func: function to calculate the value of the field.
-        :type encoder: :class:`~kitty.model.low_levele.encoder.BitFieldEncoder`
-        :param encoder: encoder for the field (default: ENC_INT_DEFAULT)
-        :param fuzzable: is container fuzzable
-        :param name: (unique) name of the container (default: None)
-        '''
-        if correction:
-            if not isinstance(correction, types.FunctionType):
-                if not isinstance(correction, types.IntType):
-                    raise KittyException('correction must be int, function or None!')
-        self._correction = correction
-        bit_field = BitField(value=0, length=length, encoder=encoder)
-        super(FieldIntProperty, self).__init__(depends_on=depends_on, bit_field=bit_field, calc_func=None, fuzzable=fuzzable, name=name)
-
-    def _calculate_value(self):
-        calculated = self._calculate(self._field)
-        if isinstance(self._correction, types.FunctionType):
-            calculated = self._correction(calculated)
-        elif isinstance(self._correction, types.IntType):
-            calculated += self._correction
-        return calculated
-
-
-class ElementCount(FieldIntProperty):
-    '''
-    Number of elements inside another field.
-    The value depends on the number of fields in the field it depends on.
-    '''
-    def _calculate(self, field):
-        return len(field.get_rendered_fields(RenderContext(self)))
-
-
-class IndexOf(FieldIntProperty):
-    '''
-    Index of a field in its container.
-
-    Edge case behavior:
-
-        - If field has no encloser - return 0
-        - If field is not rendered - return len(rendered element list) as index/
-    '''
-
-    def _calculate(self, field):
-        '''
-        We want to avoid trouble, so if the field is not enclosed by any other field,
-        we just return 0.
-        '''
-        encloser = field._enclosing
-        if encloser:
-            rendered = encloser.get_rendered_fields(RenderContext(self))
-            if field not in rendered:
-                value = len(rendered)
-            else:
-                value = rendered.index(field)
-        else:
-            value = 0
-        return value
-
-
-class Checksum(CalculatedInt):
-    '''
-    Checksum of another container.
-    '''
-    _algos = {
-        'adler32': zlib.adler32,
-        'crc32': zlib.crc32,
-    }
-
-    def __init__(self, depends_on, length, algorithm='crc32', encoder=ENC_INT_DEFAULT, fuzzable=False, name=None):
-        '''
-        :param depends_on: (name of) field to be checksummed
-        :param length: length of the checksum field (in bits)
-        :param algorithm: checksum algorithm name (from Checksum._algos) or a function to calculate the value of the field. func(Bits) -> int
-        :type encoder: :class:`~kitty.model.low_levele.encoder.BitFieldEncoder`
-        :param encoder: encoder for the field (default: ENC_INT_DEFAULT)
-        :param fuzzable: is field fuzzable (default: False)
-        :param name: (unique) name of the field (default: None)
-        '''
-        if algorithm in Checksum._algos:
-            func = Checksum._algos[algorithm]
-        else:
-            try:
-                res = algorithm(empty_bits)
-                kassert.is_of_types(res, types.IntType)
-                func = algorithm
-            except:
-                raise KittyException('algorithm should be a func(str)->int or one of the strings %s' % (Checksum._algos.keys(),))
-
-        def calc_func(x):
-            return func(x.bytes) & 0xffffffff
-
-        bit_field = BitField(value=0, length=length, encoder=encoder)
-        super(Checksum, self).__init__(depends_on=depends_on, bit_field=bit_field, calc_func=calc_func, fuzzable=fuzzable, name=name)
-
-
-class Size(CalculatedInt):
-    '''
-    Size of another container. Calculated in each render call
-
-    .. note::
-
-        In most cases you can use the function :func:`~kitty.model.low_level.aliases.SizeInBytes`
-        instead, which receives the same arguments except of `calc_func`
-    '''
-
-    def __init__(self, sized_field, length, calc_func=lambda x: len(x) / 8, encoder=ENC_INT_DEFAULT, fuzzable=False, name=None):
-        '''
-        :param sized_field: (name of) field to be sized
-        :param length: length of the size field (in bits)
-        :param calc_func: function to calculate the value of the field. func(bits) -> int (default: length in bytes)
-        :type encoder: :class:`~kitty.model.low_levele.encoder.BitFieldEncoder`
-        :param encoder: encoder for the field (default: ENC_INT_DEFAULT)
-        :param fuzzable: is field fuzzable (default: False)
-        :param name: (unique) name of the field (default: None)
-        '''
-        bit_field = BitField(value=0, length=length, encoder=encoder)
-        super(Size, self).__init__(depends_on=sized_field, bit_field=bit_field, calc_func=calc_func, fuzzable=fuzzable, name=name)
